Move actions into vet360 directory
