+++ src/platform/user/profile/reducers/index.js
@@ -13,7 +13,6 @@ import {
   CREATE_MHV_ACCOUNT_FAILURE,
   CREATE_MHV_ACCOUNT_SUCCESS,
   REMOVING_SAVED_FORM_SUCCESS,
-  UPDATE_VET360_PROFILE_FIELD,
 } from '../actions';
 
 const MAX_POLL_TIMES = 10;
@@ -61,13 +60,6 @@ function profileInformation(state = initialState, action) {
       const newState = mapRawUserDataToState(action.payload);
       return merge(state, newState);
     }
-
-    case UPDATE_VET360_PROFILE_FIELD:
-      return set(
-        ['vet360', action.fieldName],
-        merge(state.vet360[action.fieldName], action.newValue),
-        state,
-      );
 
     case PROFILE_LOADING_FINISHED:
     case UPDATE_LOGGEDIN_STATUS:
